Internalizando Posicao em CondicaoVitoria
