--- src/main/java/jogodavelha/Posicao.java
@@ -1,14 +1,0 @@
-package jogodavelha;
-
-public class Posicao {
-
-	public final int linha;
-	public final int coluna;
-	
-	public Posicao(int linha, int coluna) {
-		super();
-		this.linha = linha;
-		this.coluna = coluna;
-	}
-	
-}
